Refactor share logic into useShareSavedSearch hook and fix timer leak
- Extract duplicated handleShare + copiedHash state into shared hook
- Store timeout ID in ref, clear on unmount and before re-trigger
- Fix 'URL Copied' capitalization to match style guide
